refactoring class setup
