+++ README.md
@@ -65,7 +65,7 @@ rails g loggerator:log -a myapp
 Or you can set it manually or through configuration in the generated initialization file `config/initializers/log.rb`. For example:
 
 ```ruby
-Loggerator::Log.default_context = { app: Config.app_name }
+Loggerator.config.default_context = { app: Config.app_name }
 ```
 
 #### Custom Log Subscriber
@@ -79,7 +79,7 @@ app=example_app method=GET path=/ format=*/* controller=main action=index status
 You can disable this, using the default logger by adding or updating `config/initializers/log.rb`, to contain the following.
 
 ```
-LoggeratorRails.use_default_subscribers!
+Loggerator.config.rails_default_subscribers = true
 ```
 
 ### Metrics Integration
@@ -123,7 +123,7 @@ rails g loggerator:log -a myapp
 Or you can set it manually or through configuration in the generated initialization file `config/initializers/log.rb`. For example:
 
 ```ruby
-Loggerator::Metrics.name = Config.app_name
+Loggerator.config.metrics_app_name = Config.app_name
 ```
 
 ### Namespaces
@@ -148,8 +148,10 @@ Logs are sent to `stdout` and `stderr` as a stream to follow the [12factor](http
 Assume that the following examples have the following configuration:
 
 ```ruby
-Loggerator::Log.default_context = { app: 'myapp' }
-Loggerator::Metrics.name        = 'myapp'
+Loggerator.config do |c|
+  c.default_context = { app: "myapp" },
+  c.metrics_app_name = "myapp"
+end
 ```
 
 ### Log Helpers
@@ -380,10 +382,26 @@ Foo.new.test
 
 There are a few configuration options that can be setup in your initializer.
 
-For any key-values you want on every log message.  With the generated intializer, the app name is the default, but any data can be added to this hash.
-
 ```ruby
-Loggerator::Log.default_context = { app: "myapp", env: Rails.env }
+Loggerator.config do |c|
+
+  # Set loggerator's default context. These are the key/value pairs
+  # defining your application, which are prepended to every log line.
+  c.default_context = { app: "myapp", env: Rails.env }
+
+  # Set loggerator's metrics name. This is the name to be included as
+  # part of the metric key when emitting metrics.
+  c.metrics_app_name = Config.app_name
+
+  # Requiring 'loggerator/rails' automatically overrides Rails' log subscribers
+  # for controller handling with it's own log subscriber.
+  #
+  # In case you may need to disable this functionality, the following is
+  # a simple method for turning this off, causing the default logging to be
+  # modified.
+  c.rails_default_subscribers = true
+
+end
 ```
 
 You can also provide different streams to use instead of `stdin` and `stderr`
@@ -393,12 +411,6 @@ Here's a contrived example:
 $log = StringIO.new
 Loggerator::Log.stdout = $log
 Loggerator::Log.stderr = $log
-```
-
-And you can set the name to use for all your metric logs
-
-```ruby
-Loggerator::Metrics.name = Config.app_name
 ```
 
 ## Testing
